feat: abstract GpuController
